+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unpublished]
 ### Added
 - Added `~jot` command - Displays how much XP you'd gain from Jack of Trades based on type and skill level.
+- Added `~statues` command - Displays how much XP a specific user would gain in Construction, Prayer, and Slayer, from God statues.
 
 ### Fixed
 - Fixed price from showing up arrow when there was no change.